Run script mode is working.  Starting to hammer on the sleep and pause modes.

+++ src/testing/unit-test/cmd/TEST_A_CLI_Command.cpp
@@ -32,14 +32,14 @@ TEST( A_CLI_Command, Parameterized_Constructor_Mode )
     ASSERT_EQ( command03.Get_Description(), "");
     ASSERT_EQ( command04.Get_Description(), "");
     
-    ASSERT_EQ( command01.Get_Command_Name_List().size(), 0);
-    ASSERT_EQ( command02.Get_Command_Name_List().size(), 0);
-    ASSERT_EQ( command03.Get_Command_Name_List().size(), 0);
-    ASSERT_EQ( command04.Get_Command_Name_List().size(), 0);
+    ASSERT_EQ( (int)command01.Get_Command_Name_List().size(), 0);
+    ASSERT_EQ( (int)command02.Get_Command_Name_List().size(), 0);
+    ASSERT_EQ( (int)command03.Get_Command_Name_List().size(), 0);
+    ASSERT_EQ( (int)command04.Get_Command_Name_List().size(), 0);
 
     // Check the name list
     command01.Add_Name("Hello");
-    ASSERT_EQ( command01.Get_Command_Name_List().size(), 1);
+    ASSERT_EQ( (int)command01.Get_Command_Name_List().size(), 1);
     ASSERT_EQ( command01.Get_Command_Name_List()[0], "Hello");
 
     // Test == Operators
@@ -165,10 +165,10 @@ TEST( A_CLI_Command, Parameterized_Constructor_Full )
     ASSERT_EQ( cli_command04.Get_Description(), description04 );
 
     // Check the command list
-    ASSERT_EQ( cli_command01.Get_Command_Name_List().size(), 1 );
-    ASSERT_EQ( cli_command02.Get_Command_Name_List().size(), 1 );
-    ASSERT_EQ( cli_command03.Get_Command_Name_List().size(), 1 );
-    ASSERT_EQ( cli_command04.Get_Command_Name_List().size(), 2 );
+    ASSERT_EQ( (int)cli_command01.Get_Command_Name_List().size(), 1 );
+    ASSERT_EQ( (int)cli_command02.Get_Command_Name_List().size(), 1 );
+    ASSERT_EQ( (int)cli_command03.Get_Command_Name_List().size(), 1 );
+    ASSERT_EQ( (int)cli_command04.Get_Command_Name_List().size(), 2 );
 
     // Check the command list values
     ASSERT_EQ( cli_command01.Get_Command_Name_List()[0], "hello");
